removed full toggle / moved toggle from ui to PreviewCode
