+++ src/app/docs/components/toggle/toggle/page.mdx
@@ -1,5 +1,5 @@
 import PreviewCodeTabs from '@/components/CodePreview/page'
-import { Toggle } from '@/ui/toggle'
+import { Toggle } from '@/components/PreviewCode/Toggle'
 import Link from 'next/link'
 import { ChevronLeft } from 'lucide-react'
